refactor: 필요없는 test 제거
