+++ src/test/java/chess/domain/piece/PieceTest.java
@@ -7,7 +7,6 @@ import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 
-import static chess.domain.piece.position.PiecePosition.of;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
@@ -20,7 +19,7 @@ class PieceTest {
     @Test
     void 위치와_색상을_가지고_생성된다() {
         // when & then
-        assertDoesNotThrow(() -> new Piece(Color.WHITE, of(1, 'a')) {
+        assertDoesNotThrow(() -> new Piece(Color.WHITE, PiecePosition.of(1, 'a')) {
 
             @Override
             protected void validateMovable(final Path path) {
@@ -31,7 +30,7 @@ class PieceTest {
     @Test
     void clone_할_수_있다() {
         // given
-        final MyPiece myPiece = new MyPiece(Color.BLACK, of(1, 'a'));
+        final MyPiece myPiece = new MyPiece(Color.BLACK, PiecePosition.of(1, 'a'));
 
         // when
         final Piece clone = myPiece.clone();
@@ -53,16 +52,16 @@ class PieceTest {
     @Test
     void 경유지탐색_시_같은_위치면_예외() {
         // given
-        Piece myPiece = new MyPiece(Color.BLACK, of(1, 'a'));
+        Piece myPiece = new MyPiece(Color.BLACK, PiecePosition.of(1, 'a'));
         // when & then
-        assertThatThrownBy(() -> myPiece.waypoints(of(1, 'a')))
+        assertThatThrownBy(() -> myPiece.waypoints(PiecePosition.of(1, 'a')))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
     void 경유지탐색_시_도달불가능하면_오류() {
         // given
-        Piece myPiece = new Piece(Color.BLACK, of(1, 'a')) {
+        Piece myPiece = new Piece(Color.BLACK, PiecePosition.of(1, 'a')) {
             @Override
             protected void validateMovable(final Path path) {
                 throw new IllegalArgumentException("도달 불가");
@@ -70,7 +69,7 @@ class PieceTest {
         };
 
         // when & then
-        assertThatThrownBy(() -> myPiece.waypoints(of(1, 'b')))
+        assertThatThrownBy(() -> myPiece.waypoints(PiecePosition.of(1, 'b')))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("도달 불가");
     }
